“7/28１回目ファイルの追加”

+++ mock-cace-furima-app/src/database/migrations/2014_10_12_000000_create_users_table.php
@@ -31,6 +31,9 @@ class CreateUsersTable extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('users');
+        if (app()->environment('local')) 
+        {
+            Schema::dropIfExists('users');
+        }
     }
 }
